requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.3
django-extensions==3.2.3
pillow==11.0.0
qrcode==8.0
sqlparse==0.5.1
tzdata==2024.2
